Restructure interop tests
