+++ unit-tests/src/test/scala/scala/scalanative/native/CFuncPtrBoxingSuite.scala
@@ -1,0 +1,4 @@
+package scala.scalanative
+package native
+
+object CFuncPtrBoxingSuite extends tests.Suite {}
