+++ unit-tests/src/test/scala/scala/scalanative/native/CFuncPtrOpsSuite.scala
@@ -1,0 +1,17 @@
+package scala.scalanative
+package native
+
+import scalanative.libc._
+
+object CFuncPtrOpsSuite extends tests.Suite {
+
+  def randFunc = new CFuncPtr0[CInt] {
+    def apply(): CInt = stdlib.rand()
+  }
+
+  test("CFuncPtr cast and call with given signature") {
+    assertThrows[ClassCastException] {
+      randFunc.asInstanceOf[CFuncPtr1[Int, Int]] // wrong arity
+    }
+  }
+}
