+++ unit-tests/src/test/scala/scala/scalanative/native/CStructBoxingSuite.scala
@@ -1,0 +1,4 @@
+package scala.scalanative
+package native
+
+object CStructBoxingSuite extends tests.Suite {}
