+++ unit-tests/src/test/scala/scala/scalanative/native/CStructOpsSuite.scala
@@ -1,0 +1,4 @@
+package scala.scalanative
+package native
+
+object CStructOpsSuite extends tests.Suite {}
